+++ clients/client-python/gravitino/__init__.py
@@ -28,3 +28,4 @@ from gravitino.client.gravitino_admin_client import GravitinoAdminClient
 from gravitino.client.gravitino_metalake import GravitinoMetalake
 from gravitino.name_identifier import NameIdentifier
 from gravitino.filesystem import gvfs
+from gravitino.namespace import Namespace
