+++ clients/client-python/gravitino/api/rel/indexes/__init__.py
@@ -16,3 +16,4 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 """
+from gravitino.api.rel.indexes.index import Index
