+++ clients/client-python/gravitino/api/rel/types/__init__.py
@@ -16,3 +16,5 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 """
+
+from gravitino.api.rel.types.type import Type
