+++ clients/client-python/gravitino/dto/responses/__init__.py
@@ -16,3 +16,5 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 """
+from gravitino.dto.responses.base_response import BaseResponse
+from gravitino.dto.responses.table_response import TableResponse
