+++ clients/client-python/gravitino/exceptions/__init__.py
@@ -16,3 +16,10 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 """
+
+from gravitino.exceptions.handlers import TABLE_ERROR_HANDLER
+from gravitino.exceptions.base import (
+    IllegalNamespaceException,
+    IllegalNameIdentifierException,
+    IllegalArugmentException,
+)
