+++ clients/client-python/gravitino/exceptions/handlers/__init__.py
@@ -16,3 +16,4 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 """
+from gravitino.exceptions.handlers.table_error_handler import TABLE_ERROR_HANDLER
