+++ clients/client-python/gravitino/rest/__init__.py
@@ -16,3 +16,4 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 """
+from gravitino.rest.rest_utils import encode_string
